# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "weblav",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/backup/restore": {
      "post": {
        "summary": "受け取ったバックアップで、今のデータを置き換える。全員のセッションを消すので、この要求の送り手もログインし直す。",
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreBackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "戻した"
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin以外",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "受け取ったものが無い・期限切れ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "アーカイブの走査中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "バックアップが壊れている",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ValidationDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "API エラーの機械可読なコード。frontend はこの値で表示文言を選ぶ。",
        "enum": [
          "not_found",
          "database_unavailable",
          "unauthorized",
          "forbidden",
          "invalid_credentials",
          "incorrect_current_password",
          "invalid_recovery_code",
          "too_many_requests",
          "invalid_request_body",
          "file_too_large",
          "internal_error",
          "conflict",
          "config_unreadable",
          "free_limit_reached",
          "account_server_unavailable",
          "pro_code_invalid",
          "pro_clock_behind"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通envelope。`code` が機械可読な契約(frontendはこちらでi18nする)、\n`message` は英語のデバッグ用の説明文で、画面には出さない。\n\n汎用の文言では直しようがない422には `detail` を添える。こちらも機械可読な契約で、\n画面が値を差し込んで文言を組み立てる (→ `api::error_detail`)。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "FreeLimitTarget": {
        "type": "string",
        "description": "上限のある対象。",
        "enum": [
          "admin",
          "user",
          "archive",
          "folder",
          "file",
          "link"
        ]
      },
...
      "RestoreBackupRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "受け取ったときに返した id。"
          }
        }
      },
...
      "ValidationDetail": {
        "oneOf": [
          {
            "type": "object",
            "description": "グループは「本人のみ」にできない (→ docs/access.md「公開範囲」)。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "groupCannotBePrivate"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "軸名にプレースホルダーの記号 (`{` `}`) を含んでいる。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "axisNameHasBraces"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "同じ名前の軸が既にある。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "axisNameTaken"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "予約プレースホルダーと同じ名前は軸に使えない (→ docs/archive.md「表示タイトル」)。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "axisNameReserved"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "表示タイトルのテンプレートが、存在しない軸を参照している。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "templateUnknownAxis"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "外部の AI の答えを取り込むとき、この軸の定義が決まりに合わない (階層の番号・値の空や重複など)。\nどの決まりかまでは分けない。答えを直すのは AI に頼み直すことになり、軸が分かれば足りるため。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "importAxisInvalid"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "表示タイトルのテンプレートで使われている軸は削除できない。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "axisInUseByTemplate"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "同じ名前のユーザーが既にある。大文字小文字だけの違いも同じ名前とみなす。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "usernameTaken"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "最後の管理者は降格・削除できない (→ docs/access.md「admin の最後の1人」)。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "lastAdmin"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "同じ名前の「公開できるフォルダー」が既にある (→ docs/folders.md「公開できるフォルダー」)。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "rootNameTaken"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "そのフォルダーは「公開できるフォルダー」に登録済み (→ docs/folders.md「公開できるフォルダー」)。",
            "required": [
              "name",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "rootAlreadyRegistered"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "WebLAV 自身の置き場 (設定とデータ) の中は、公開できるフォルダーにできない (→ docs/folders.md「公開できるフォルダー」)。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "rootInsideOwnDirs"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "受け取ったファイルが WebLAV のバックアップではない・壊れている (→ docs/access.md「バックアップとリストア」)。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "backupInvalid"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "受け取ったバックアップが、今より新しい版で作られている (→ docs/access.md「バックアップとリストア」)。",
            "required": [
              "appVersion",
              "kind"
            ],
            "properties": {
              "appVersion": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "backupTooNew"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "再スキャンで、アーカイブの登録先のフォルダーが見つからない (→ docs/archive.md「スキャン」)。",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "archiveFolderMissing"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "再スキャンで、読めない場所があった (→ docs/archive.md「スキャン」)。\n`path` は最初に読めなかった場所の、登録先からの相対パス。登録先そのものなら空文字。",
            "required": [
              "path",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "archiveFolderUnreadable"
                ]
              },
              "path": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Free の上限に当たった (→ docs/pro.md「上限を数えて止める」)。409 に付ける。",
            "required": [
              "target",
              "limit",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "freeLimit"
                ]
              },
              "limit": {
                "type": "integer",
                "format": "int64"
              },
              "target": {
                "$ref": "#/components/schemas/FreeLimitTarget"
              }
            }
          }
        ],
        "description": "422 の内訳。`kind` で場合分けし、文言に差し込む値を併せて渡す。"
      },
...
    }
  }
}